liblog: Silence logs from proprietary camera libraries (AEC_PORT)
Our proprietary camera libraries are spamming logs very hard, especially when video HDR mode is enabled. The debug spam with video HDR mode is so great that the log gets overflowed and the camera is subsequently killed.

Silence all AEC_PORT logs to fix the camera crashes. These logs are not useful for debugging anyway.

Change-Id: I21c9e172688c116d0f6d1419125f2b82be6bbdce